tests/wpt/web-platform-tests/css/css-flexbox/intrinsic-size/col-wrap-010.html

<!DOCTYPE html>
<link rel="author" title="David Grogan" href="mailto:dgrogan@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#intrinsic-sizes">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>

<meta name="assert"
  content="Item with unresolvable width:100% contributes to max-content size via its content contributions." />

<style>
  .grandchild {
    float: left;
    width: 50px;
  }

</style>

<div
  style="display: flex; flex-flow: column wrap; width: max-content; height: 100px; background: green;"
  data-expected-width="100">
  <div style="width: 100%; height: 50px; min-height: 0px;"
    data-expected-width="100">
    <div class="grandchild"></div>
    <div class="grandchild"></div>
  </div>
</div>


<script>
  checkLayout('body > div');
</script>
